Ref: Major refactor
