+++ src/demo/components/index.ts
@@ -1,0 +1,5 @@
+export { default as Header } from './Header';
+export { default as Installation } from './Installation';
+export { default as ResponsiveExamples } from './ResponsiveExamples';
+export { default as LineChartExamples } from './LineChartExamples';
+export { default as BarChartExamples } from './BarChartExamples';
