+++ src/lib/BarChart/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @file index.ts
+ * @description Export file for the BarChart component
+ */
+
+export * from './BarChart';
+export * from './types';
