+++ src/lib/LineChart/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @file index.ts
+ * @description Export file for the LineChart component
+ */
+
+export * from './LineChart';
+export * from './types';
