+++ src/lib/index.ts
@@ -1,4 +1,19 @@
-export { BarChart } from './BarChart';
-export type { BarChartProps } from './BarChart';
-export { LineChart } from './LineChart';
-export type { LineChartProps } from './LineChart'; 
+/**
+ * @file index.ts
+ * @description Main export file for the chart library
+ */
+
+// Export components
+export { BarChart } from './BarChart/BarChart';
+export type { BarChartProps, BarChartData } from './BarChart/types';
+
+export { LineChart } from './LineChart/LineChart';
+export type { LineChartProps, LineChartSeries, LineChartDataPoint } from './LineChart/types';
+
+// Export hooks for advanced usage
+export { useChartDimensions } from './hooks/useChartDimensions';
+export { useTooltip } from './hooks/useTooltip';
+
+// Export utilities
+export { createGradient, addGridLines, addLegend } from './utils/chartUtils';
+export type { Margin } from './utils/chartUtils';
